Correction typage

+++ C/math.c
@@ -1,6 +1,6 @@
 #include "math.h"
 
-int maximum (int n1, int n2 )
+float maximum (float n1, float n2 )
 {
    return (n1 >= n2)? n1 : n2;
 }
